Added an app.yaml for appengine deployment.
We can then deploy this to beta.bigoquiz.com like so, after doing "ng
build":
gcloud app deploy

This works because there is a wildcard domain mapping (*.bigoquiz.com)
in the AppEngine project settings. The subdomain name then maps to
any matching service name (this one is called "beta" in the app.yaml).