UserStatus: Show the Google Sign-In button.
Using the "custom Google Sign-In button" technique: Adding a div
and calling Google's javascript to change it:
https://developers.google.com/identity/sign-in/web/build-button

We cannot use the simple "automatically rendered" button,
https://developers.google.com/identity/sign-in/web/sign-in ,
because the html in the Angular component is not rendered until
after Google' javascript tries to find the div.

This currently uses the oauth credentials client id for
beta.bigoquiz.com. This is apparently not meant to be secret.