UserHistorySectionsComponent: problem question links: Don't always specify section.
Only specify a section ID i the link if the page is currently using a section
for next questions , and if that section is the same as the question's section.

This is something that gwt-bigoquiz currently gets wrong: It takes you to the
question, which might be in section foo, while saying that its showing
questions from section bar.