UserHistoryService: Properly parse QuestionHistories.
Use for of instead of for in, because they are in a simple JSON array,
not a mapping. Before, numerical array indices were being used as
question IDs, which were then used as keys in the map, causing
questions to be in the map twice when the same question was answered.