app.yaml: Move skipfiles to .gcloudignore
And add some more files to ignore.

As per:

ERROR: (gcloud.app.deploy) skip_files cannot be used with the [python313] runtime. Ignore patterns are instead expressed in a .gcloudignore file. For information on the format and syntax of .gcloudignore files, see https://cloud.google.com/sdk/gcloud/reference/topic/gcloudignore.

As per
https://docs.cloud.google.com/appengine/docs/flexible/reference/app-yaml?tab=python
this does not add app.yaml to the .gicloudignore file