app.component test: Import (standalone) AppComponent
And remove it from the declarations.

To avoid this error with (as seen in the GitHub Action workflow):

"
Failed: Unexpected "AppComponent" found in the "declarations" array of
the "TestBed.configureTestingModule" call, "AppComponent" is marked as
standalone and can't be declared in any NgModule - did you intend to
import it instead (by adding it to the "imports" array)?
"